requirements.txt
Django (1.11.5)
djangorestframework (3.6.4)
lxml (3.8.0)
pip (8.1.1)
pytz (2017.2)
setuptools (20.10.1)
